Create file
The code defines two structures, Node and List, to create a linked list data structure. The Node structure holds a pointer to the data stored in the node, and a pointer to the next node in the list. The List structure holds a pointer to the head node of the list and an integer representing the size of the data stored in each node.

The linked list is initialized using the list_init function, which sets the head pointer of the List structure to NULL and the data_size to the specified size.

The list_insert function is used to insert new nodes into the list. The function creates a new Node structure, allocates memory for the node's data, sets the next node to the current head of the list, and copies the data into the node's data. The head of the list is then set to the new node.

The list_reverse function reverses the order of the nodes in the linked list. The function uses three pointers, prev, current, and next, to traverse the list and reverse the order of the nodes.

The list_print function allows for printing the data in each node of the linked list. The function takes a pointer to a List structure and a pointer to a print function, which is used to print the data stored in each node. The function traverses the list, calling the print function on the data of each node.

The int_print function is used to print the data as integers.

Finally, the code demonstrates the use of the linked list by initializing a list of integers, inserting 5 elements into the list, printing the original list, reversing the list, and printing the reversed list.

This code implements a basic linked list in C, showcasing knowledge of pointers, dynamic memory allocation, and linked list concepts.