Create Linked List with Variable Average
This program defines a linked list data structure to store a variable number of integers. It uses dynamic memory allocation to create new Nodes in the list and pointers to traverse and manipulate the list. It also uses the <stdarg.h> library to compute the average of a variable number of integers.

In the main function, the program creates a linked list with four integers, and then computes the average of those integers using the average function. Finally, the program prints the contents of the list and the computed average..